requirements.txt
pycaw
comtypes
winrt-runtime
PyQt6